Wave generator generating enemies according to Points 	- waveGenerator is now generating waves.
Known Bugs: 
	- just spawned enemy size is Larger than has to be
	- Projectile passes through some enemies
	- Sometimes enemies spawn outside the map(an if problem), sometimes
inside collision(an if problem)...

+++ SwampMonster/src/com/me/swampmonster/models/Wave.java
@@ -6,5 +6,5 @@ public class Wave {
 	public Stack<Enemy> enemies;
 	public int enemiesOnBattleField;
 	public int pendingPeriod;
-	public int rate;
+	public float rate;
 }
